modem: always check for a SIM in CDMA-only modems
This patch fixes problems we have with ModemManager's EVDO-CDMA
capability.
It stems from the fact that a SIM is not required for CDMA.
However, we only deal with modems with SIMs.

The problems:
1. ModemManager doesn't check for a SIM in CDMA-only modems.
2. ModemManager sets CDMA mode for 3GPP/CDMA modems with no SIM,
   because it assumes that has the best chance of working.
3. ModemManager doesn't fail SIM-less CDMA-mode modems because it
   thinks they will still work. In this case, a SIM is created with
   unknown IMSI, ICCID, etc.

Fixing Problem 1 is necessary for an external USB modem (not sure
which) that is CDMA-only but has a SIM that is used.

Fixing Problems 2 & 3 is necessary to prevent ghost SIMs in e.g.
EVDO-compatible MC7354 (SKU 1102552) modems.