qmi: DHCP bearer default
With the old 4.4 kernel we only ever ran DHCP on QMI modems because the
kernel did not support the ioctl to query raw-ip mode.

With the new kernel the ioctl is now supported and w4e are using static
for
Sierra modems.  Since this is a major change to what we have used in the
past we are going to temporarily turn this off in ModemManager to ensure
less change in the next release and less chance of a stuff up.