mbim: support multiple VLAN sessions
These changes support using multiple bearers on MBIM modems:
  1. The bearer "number" parameter is the requested MBIM Session ID.
  2. The bearer grabs the net port associated with the MBIM Session ID.

E.g.:
  Session ID 0 -> wwan0
  Session ID 1 -> wwan0.1
  Session ID 2 -> wwan0.2
  etc...

If no bearer "number" is provided, the search for the first unused session ID
will occur as before.

Unfortunately, the mapping of PDP profile index to Session ID is based on the
assumption that the modem has been pre-configured to do this.
E.g. for Telit LE910-NA V2 in AT&T mode, if you run
  AT#MBIMCFG=1,2,3,4,5,6,7,8,9,10,11,12,13,14,15
Then Session ID 0 is mapped to PDP profile 1, Session ID 1 is mapped to PDP
profile 2, etc.
The default #MBIMCFG value is usually "14" (for AT&T) or "3" (for Verizon),
which means that only Session ID 0 works by default as other sessions are not
mapped to a PDP profile.