bearer: reinstate "number" attribute
This reverts the parts of ModemManager commit
55c3026643ee6c75c0a68171abd884edf035b9ba that removed the "number" from the
Bearer Properties. More specifically, that commit left the API for the
"number" in place but made it non-functional.

The bearer "number" is used by patches for nominating PDP context ids for
bearers, which is required for dual APNs with MBIM and QMI muxing.