mbim: poll for SMS
The Telit LE910NA-V2 does not send MBIM events for received SMSs even though
ModemManager subscribes to them.
This means that SMSs were read from the modem only when it was transitioning
from disabled to enabled state.
This HACK re-runs the SMS enabling process on the signal strength poll
timer in order to receive SMSs without having to disable and then enable
the modem.

This will affect all MBIM modems that do the periodic poll, which right now
is only the Telit MBIM modems we use.