kernel-device-udev: fix memory usage after free'd up problem
When kernel_device_get_driver() function returns the driver name of the
parent device (return value of g_udev_device_get_parent() /
get_parent()), that string is stored in the parent device's structure.
As the parent device is free'd up right in the function, cannot return
with the string, without duplicating it.

Cache the driver name on the first search, and return with the cached
one always.