quectel: reprobe modem on unsolicited RDY msg
There are reports of the Quectel EG25G performing an internal reset for
unknown reasons, which causes the modem to reset without dropping off
the USB bus. ModemManager does not notice that the modem has reset,
and continues with stale QMI sessions that no longer work.

ModemManager debug logs from when this occur look like this:

    <debug> [000253.216561] [modem1/ttyUSB3/at] <-- '<CR><LF>RDY<CR><LF>'
    <debug> [000253.217746] [modem1/ttyUSB2/at] <-- '<CR><LF>RDY<CR><LF>'
    <debug> [000253.244104] [modem1/ttyUSB2/at] <-- '<CR><LF>+CPIN: READY<CR><LF>'
    <debug> [000253.266475] [modem1/ttyUSB2/at] <-- '<CR><LF>+QUSIM: 1<CR><LF>'
    <debug> [000253.266664] [modem1] (quectel) checking SIM swap
    <debug> [000253.266745] [modem1] checking if SIM was swapped...
    <debug> [000253.266909] [modem1/sim1] loading SIM identifier...
    ...
    transaction 0x405 aborted, but message is not abortable
    <warn>  [000261.008737] [modem1/bearer1] reloading stats failed: QMI operation failed: Transaction timed out

It can be detected by handling the unsolicited "RDY" message on the AT
ports that is sent when the modem starts up. Making ModemManager reprobe
the modem gets it working again.

This patch hijacks the Quectel SIM hot swap setup which already sets an
unsolicited message handler, for lack of a more appropriate place
currently.
Adding the message handler happens late enough after the initial probe
(after e.g. a hardware reset), so that any first RDY does not trigger the
reprobe.