Author: David Leonard Date: 2020-08-28 11:51:20 Revision: 442e5aceabc5e9c5704fbf50443efd0278a8bd93 Comment: ModemManager: allow fallback from CDMA to LTE
Remove a restriction that fails a CDMA network connection if an APN
had been specified.
Although an APN is not needed for CDMA EV-DO, some carriers like
US Cellular need to be able to fall back from LTE to those types,
and an APN will have been specified.

OG-8176 OG-4741