Author: David McCullough Date: 2020-09-14 11:16:47 Revision: 441742de2b22fe9c87e157db31a32af3a5f0a95b Comment: ModemManager: fix various build issues
The aclocal-1.15 patch caused build failure on systems with only aclocal-1.16

As with libqmi we need to force autogen/autoreconf to ensure all versions are
correct. Unfortunately doing this introduced a failure because the gettext tool
version is forced to 0.17, so remove the check, we will learn soon enough
if it will not work.